requirements.txt
py-dom-xpath
requests
